refact: 소켓 핸들러 Dispatch 분기 로직 분리
